add-director
Исправление ошибок для postman по различным задачам

+++ src/main/java/ru/yandex/practicum/filmorate/repository/impl/JdbcReviewRepository.java
@@ -1,7 +1,6 @@
 package ru.yandex.practicum.filmorate.repository.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.aspectj.weaver.reflect.IReflectionWorld;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
@@ -10,13 +9,13 @@ import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.model.Review;
+import ru.yandex.practicum.filmorate.model.UserEvent;
 import ru.yandex.practicum.filmorate.repository.ReviewRepository;
 import ru.yandex.practicum.filmorate.repository.mappers.ReviewRowMapper;
+import ru.yandex.practicum.filmorate.service.EventService;
 
 import java.sql.PreparedStatement;
-import java.sql.Types;
 import java.util.List;
-import java.util.Objects;
 import java.util.Optional;
 
 @Repository
@@ -25,6 +24,8 @@ public class JdbcReviewRepository implements ReviewRepository {
     private final NamedParameterJdbcOperations jdbc;
     private final JdbcTemplate jdbcTemplate;
     private final ReviewRowMapper mapper;
+
+    private final EventService eventService;
 
     @Override
     public Review save(Review review) {
@@ -73,8 +74,7 @@ public class JdbcReviewRepository implements ReviewRepository {
 
     @Override
     public Review update(Review review) {
-        String query = "UPDATE reviews SET content = ?, is_positive = ?, " +
-                "user_id = ?, film_id = ?, useful = ? " +
+        String query = "UPDATE reviews SET content = ?, is_positive = ? " +
                 "WHERE review_id = ?";
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
@@ -82,10 +82,7 @@ public class JdbcReviewRepository implements ReviewRepository {
             PreparedStatement statement = connection.prepareStatement(query);
             statement.setString(1, review.getContent());
             statement.setBoolean(2, review.getIsPositive());
-            statement.setLong(3, review.getUserId());
-            statement.setLong(4, review.getFilmId());
-            statement.setLong(5, review.getUseful() != null ? review.getUseful() : 0);
-            statement.setLong(6, review.getReviewId());
+            statement.setLong(3, review.getReviewId());
             return statement;
         }, keyHolder);
 
@@ -122,7 +119,7 @@ public class JdbcReviewRepository implements ReviewRepository {
         String query = "SELECT review_id, content, is_positive, user_id, film_id, useful " +
                 "FROM reviews " +
                 (filmId.isPresent() ? "WHERE film_id = :filmId " : "") +
-                "ORDER BY useful DESC " +
+                "ORDER BY USEFUL DESC " +
                 "LIMIT :count";
 
         MapSqlParameterSource params = new MapSqlParameterSource()
@@ -139,26 +136,30 @@ public class JdbcReviewRepository implements ReviewRepository {
     @Override
     public void addLike(Integer reviewId, Long userId) {
         KeyHolder keyHolder = new GeneratedKeyHolder();
-        String removeDislikeQuery = "DELETE FROM review_reactions " +
-                "WHERE review_id = ? AND user_id = ? AND is_like = false";
-        int removedDislikes = jdbcTemplate.update(connection -> {
-            PreparedStatement statement = connection.prepareStatement(removeDislikeQuery);
-            statement.setLong(1, reviewId);
-            statement.setLong(2, userId);
-            return statement;
-        }, keyHolder);
+        Review review = getById(reviewId).orElseThrow();
+        if (!review.getIsPositive()) {
+            String removeDislikeQuery = "DELETE FROM review_reactions " +
+                    "WHERE review_id = ? AND user_id = ? AND is_like = false";
+            int removedDislikes = jdbcTemplate.update(connection -> {
+                PreparedStatement statement = connection.prepareStatement(removeDislikeQuery);
+                statement.setLong(1, reviewId);
+                statement.setLong(2, userId);
+                return statement;
+            }, keyHolder);
+
         /*int removedDislikes = jdbc.update(removeDislikeQuery, new MapSqlParameterSource()
                 .addValue("review_id", reviewId)
                 .addValue("user_id", userId));*/
 
 
-        if (removedDislikes > 0) {
-            String increaseUsefulQuery = "UPDATE reviews SET useful = useful + 1 WHERE review_id = ?";
-            jdbcTemplate.update(connection -> {
-                PreparedStatement statement = connection.prepareStatement(increaseUsefulQuery);
-                statement.setLong(1, reviewId);
-                return statement;
-            }, keyHolder);
+            if (removedDislikes > 0) {
+                String increaseUsefulQuery = "UPDATE reviews SET useful = useful + 1 WHERE review_id = ?";
+                jdbcTemplate.update(connection -> {
+                    PreparedStatement statement = connection.prepareStatement(increaseUsefulQuery);
+                    statement.setLong(1, reviewId);
+                    return statement;
+                }, keyHolder);
+            }
         }
 
         String insertLikeQuery = "INSERT INTO review_reactions (review_id, user_id, is_like) " +
@@ -169,6 +170,8 @@ public class JdbcReviewRepository implements ReviewRepository {
             statement.setLong(2, userId);
             return statement;
         }, keyHolder);
+
+        eventService.createEvent(userId, reviewId, UserEvent.EventType.LIKE, UserEvent.EventOperation.ADD);
 
         String increaseUsefulQuery = "UPDATE reviews SET useful = useful + 1 WHERE review_id = ?";
 
@@ -191,6 +194,8 @@ public class JdbcReviewRepository implements ReviewRepository {
             return statement;
         }, keyHolder);
 
+        eventService.createEvent(userId, reviewId, UserEvent.EventType.LIKE, UserEvent.EventOperation.REMOVE);
+
         String queryReview = "UPDATE reviews SET useful = useful - 1 WHERE review_id = ?";
         jdbcTemplate.update(connection -> {
             PreparedStatement statement = connection.prepareStatement(queryReview);
@@ -202,22 +207,25 @@ public class JdbcReviewRepository implements ReviewRepository {
     @Override
     public void addDislike(Integer reviewId, Long userId) {
         KeyHolder keyHolder = new GeneratedKeyHolder();
-        String removeLikeQuery = "DELETE FROM review_reactions " +
-                "WHERE review_id = ? AND user_id = ? AND is_like = true";
-        int removedLikes = jdbcTemplate.update(connection -> {
-            PreparedStatement statement = connection.prepareStatement(removeLikeQuery);
-            statement.setLong(1, reviewId);
-            statement.setLong(2, userId);
-            return statement;
-        }, keyHolder);
-
-        if (removedLikes > 0) {
-            String decreaseUsefulQuery = "UPDATE reviews SET useful = useful - 1 WHERE review_id = ?";
-            jdbcTemplate.update(connection -> {
-                PreparedStatement statement = connection.prepareStatement(decreaseUsefulQuery);
+        Review review = getById(reviewId).orElseThrow();
+        if (review.getIsPositive()) {
+            String removeLikeQuery = "DELETE FROM review_reactions " +
+                    "WHERE review_id = ? AND user_id = ? AND is_like = true";
+            int removedLikes = jdbcTemplate.update(connection -> {
+                PreparedStatement statement = connection.prepareStatement(removeLikeQuery);
                 statement.setLong(1, reviewId);
+                statement.setLong(2, userId);
                 return statement;
             }, keyHolder);
+
+            if (removedLikes > 0) {
+                String decreaseUsefulQuery = "UPDATE reviews SET useful = useful - 1 WHERE review_id = ?";
+                jdbcTemplate.update(connection -> {
+                    PreparedStatement statement = connection.prepareStatement(decreaseUsefulQuery);
+                    statement.setLong(1, reviewId);
+                    return statement;
+                }, keyHolder);
+            }
         }
 
         String insertDislikeQuery = "INSERT INTO review_reactions (review_id, user_id, is_like) " +
@@ -228,6 +236,8 @@ public class JdbcReviewRepository implements ReviewRepository {
             statement.setLong(2, userId);
             return statement;
         }, keyHolder);
+
+        eventService.createEvent(userId, reviewId, UserEvent.EventType.LIKE, UserEvent.EventOperation.ADD);
 
         String decreaseUsefulQuery = "UPDATE reviews SET useful = useful - 1 WHERE review_id = ?";
         jdbcTemplate.update(connection -> {
@@ -248,6 +258,8 @@ public class JdbcReviewRepository implements ReviewRepository {
             statement.setLong(2, userId);
             return statement;
         }, keyHolder);
+
+        eventService.createEvent(userId, reviewId, UserEvent.EventType.LIKE, UserEvent.EventOperation.REMOVE);
 
         String queryReview = "UPDATE reviews SET useful = useful + 1 WHERE review_id = ?";
         jdbcTemplate.update(connection -> {
